pyproject.toml
[project]
name = "pddl"
version = "0.1.0"
description = "Verificador PDDL"
readme = "README.md"
authors = [
    { name = "Fábio Macêdo Mendes", email = "fabiomacedomendes@gmail.com" },
    { name = "Lucas Martins Gabriel", email = "lucasmartinsgabriel@gmail.com" },
]
requires-python = ">=3.10"
dependencies = [
    "ipdb>=0.13.13",
 "lark-parser>=0.12.0",
 "rich>=14.0.0",
]

[project.scripts]
pddl = "pddl.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-timeout>=2.4.0",
]


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["F403", "F405"]
